+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # deepsea-components
 
+## 5.5.10
+
+### Patch Changes
+
+- Updated dependencies
+  - deepsea-tools@5.9.5
+
 ## 5.5.9
 
 ### Patch Changes